backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-socketio
python-dotenv

